Correction for both sides

+++ src/ElectroMagnBC/ElectroMagnBCAM_SM.cpp
@@ -268,7 +268,7 @@ void ElectroMagnBCAM_SM::apply( ElectroMagn *EMfields, double time_dual, Patch *
                 std::complex<double> byE = 0.;
                 // Lasers
                 if( imode==1 ) {
-                    yp[0] = patch->getDomainLocalMin( 1 ) + ( j - EMfields->oversize[1] )*dr;
+                    yp[0] = patch->getDomainLocalMin( 1 ) + ( (double)j - (double)EMfields->oversize[1] )*dr;
                     for( unsigned int ilaser=0; ilaser< vecLaser.size(); ilaser++ ) {
                         byE +=          vecLaser[ilaser]->getAmplitude0( yp, time_dual, 1+2*j, 0 )
                                         + Icpx * vecLaser[ilaser]->getAmplitude1( yp, time_dual, 1+2*j, 0 );
@@ -288,7 +288,7 @@ void ElectroMagnBCAM_SM::apply( ElectroMagn *EMfields, double time_dual, Patch *
             
                 std::complex<double> bzE = 0.;
                 if( imode==1 ) {
-                    yd[0] = patch->getDomainLocalMin( 1 ) + ( j - 0.5  - EMfields->oversize[1] )*dr;
+                    yd[0] = patch->getDomainLocalMin( 1 ) + ( (double)j - 0.5  - (double)EMfields->oversize[1] )*dr;
                     // Lasers
                     for( unsigned int ilaser=0; ilaser< vecLaser.size(); ilaser++ ) {
                         bzE +=         vecLaser[ilaser]->getAmplitude1( yd, time_dual, 2*j, 0 )
